+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./schema.js";
 export * from "./ids.js";
+export * from "./relations.js";
 export * from "./metrics.js";
 export * from "./merge.js";